+++ src/chains/evm/velas.js
@@ -4,10 +4,16 @@ module.exports = {
   NAME: 'Velas',
   RPC: 'https://explorer.velas.com/rpc',
   DECIMALS: 18,
-  GET_TXN_LIST_API: '',
+  GET_TXN_LIST_API(address, page, limit) {
+    return ``;
+  },
   ASSET_API: 'https://api.coingecko.com/api/v3/coins/velas',
-  EVM_FUNGIBLE_ASSET_DISCOVERY_API: '',
-  EVM_NON_FUNGIBLE_ASSET_DISCOVERY_API: '',
+  EVM_FUNGIBLE_ASSET_DISCOVERY_API(address, chainId) {
+    return ``
+  },
+  EVM_NON_FUNGIBLE_ASSET_DISCOVERY_API(address) {
+    return ``;
+  },
   FUNCTIONALITY_SUPPORT: {
     getTransactions: false,
     discoverFungibleAssets: false,