+++ src/utils/helper.js
@@ -1,17 +1,13 @@
 const axios = require('axios');
 
-const chains = require('../chains');
-
 const getRequest = async (url) => {
   try {
     const response = await axios.get(url);
 
     return response.data;
   } catch (err) {
-    return err;
+    throw err;
   }
 };
 
-const getApiURL = (chain, address, page, limit) => `${chains[chain].GET_TXN_LIST_API}&address=${address}&startblock=0&page=${page}&offset=${limit}&sort=asc`;
-
-module.exports = { getRequest, getApiURL };
+module.exports = { getRequest };